[vector-debug] initialize shape properties that are vector-typed by default in `insertExpr`, so overrides like `s.shape.start[0]` can work even if `s.shape.start` is not set in Style. also encode working programs as tests and break resample into separate file to avoid circular import with `EngineUtils` #515
