--- examples/vector-debug/accesspath-debug.sty
@@ -1,77 +1,0 @@
--- Colors {
---     Colors.black = rgba(0.0, 0.0, 0.0, 1.0)
---     Colors.gray = rgba(0.8, 0.8, 0.8, 1.0)
---     Colors.red = rgba(1.0, 0.0, 0.0, 1.0)
---     Colors.pink = rgba(1.0, 0.4, 0.7, 1.0)
---     Colors.yellow = rgba(1.0, 1.0, 0.0, 1.0)
---     Colors.orange = rgba(1.0, 0.6, 0.0, 1.0)
---     Colors.green = rgba(0.0, 1.0, 0.0, 1.0)
---     Colors.blue = rgba(0.0, 0.0, 1.0, 1.0)
---     Colors.sky = rgba(0.325, 0.718, 0.769, 1.0)
---     Colors.white = rgba(1.0, 1.0, 1.0, 1.0)
---     Colors.none = rgba(0.0, 0.0, 0.0, 0.0)
--- }
-
--- (2) AccessPath overrides 
-
--- (2a) Property -- DOES NOT WORK
--- Cannot read property 'tag' of undefined -- in `addPath` in `translateLine`
-
--- PROBLEM: Shape properties haven't been initialized yet here. (`start` doesn't exist as a property of the shape)
-
--- Object o {
---     shape o.shape = Line {
---         color: rgba(0., 0., 0., 1.)
---         thickness: 5.
---         rightArrowhead: true
---     }
-
---      o.shape.start[0] = 0.
--- }
-
--- (2b -- 2 examples) DID NOT WORK EARLIER, NOW BOTH WORK
--- earlier error: error: invalid varying path (in `getNum, in lookupNumericPaths`)
--- fix: add PropertyPath as a kind of AccessPath to initialize, and then initialize the shapes before the accesspaths in genOptProblem so that properties can be properly set
-
-Object o {
-    shape o.shape = Line {
-        start: (0., ?)
-        color: rgba(0., 0., 0., 1.)
-        thickness: 5.
-        rightArrowhead: true
-    }
-}
-
--- Object o {
---     shape o.shape = Line {
---           start: (?, ?)
---         color: rgba(0., 0., 0., 1.)
---         thickness: 5.
---         rightArrowhead: true
---     }
-
---     o.shape.start[0] = 0.
--- }
-
--- (2c) WORKS
-
--- Object o {
---     o.y = ?
---     shape o.shape = Line {
---         start: (0., o.y)
---         color: rgba(0., 0., 0., 1.)
---         thickness: 5.
---         rightArrowhead: true
---     }
--- }
-
--- (2d) -- Field override
--- This appears to work
-
--- Object o {
---        o.f = (?, ?)
---        o.f[0] = 0.
-
---        o.shape = Circle {}
--- }
-
