--- examples/vector-debug/test.dsl
@@ -1,1 +1,0 @@
-type Object
