--- examples/vector-debug/test.sub
@@ -1,1 +1,0 @@
-Object o
