--- examples/vector-debug/vector-debug.sty
@@ -1,53 +1,0 @@
-Colors {
-    Colors.black = rgba(0.0, 0.0, 0.0, 1.0)
-    Colors.gray = rgba(0.8, 0.8, 0.8, 1.0)
-    Colors.red = rgba(1.0, 0.0, 0.0, 1.0)
-    Colors.pink = rgba(1.0, 0.4, 0.7, 1.0)
-    Colors.yellow = rgba(1.0, 1.0, 0.0, 1.0)
-    Colors.orange = rgba(1.0, 0.6, 0.0, 1.0)
-    Colors.green = rgba(0.0, 1.0, 0.0, 1.0)
-    Colors.blue = rgba(0.0, 0.0, 1.0, 1.0)
-    Colors.sky = rgba(0.325, 0.718, 0.769, 1.0)
-    Colors.white = rgba(1.0, 1.0, 1.0, 1.0)
-    Colors.none = rgba(0.0, 0.0, 0.0, 0.0)
-}
-
--- (1) Optimizing on fields
-
--- (1a) DOES NOT WORK: (but does work after fix)
-
-Object o {
-    vec2 o.v1 = (?, ?)
-    vec2 o.v2 = (?, ?)
-
-    -- This shape is properly optimized after fix
-    shape o.shape = Line {
-        start: o.v1
-        end: o.v2
-        color: rgba(0., 0., 0., 1.)
-        thickness: 5.
-        rightArrowhead: true
-    }
-
-    ensure rightwards(o.shape)
-
-    shape o.shape1 = Line {
-        color: rgba(0., 0., 0., 1.)
-        thickness: 5.
-        rightArrowhead: true
-    }
-
-    ensure rightwards(o.shape1)
-}
-
--- (1b) DOES WORK: (before fix)
-
--- Object o {
---     shape o.shape = Line {
---         color: rgba(0., 0., 0., 1.)
---         thickness: 5.
---         rightArrowhead: true
---     }
-
---     ensure rightwards(o.shape)
--- }
